Improve the precision of coordinate transformations.

+++ common/src/main/java/io/github/reserveword/imblocker/common/gui/MinecraftMultilineEditBoxWidget.java
@@ -9,7 +9,7 @@ public interface MinecraftMultilineEditBoxWidget extends MinecraftAbstractTextIn
 		MultilineCursorInfo cursorInfo = getCursorInfo();
 		int caretX = imblocker$getContentOffsetX() + MinecraftClientAccessor.INSTANCE.getStringWidth(
 				StringUtil.getSubstring(cursorInfo.text, cursorInfo.cursorLineBeginIndex, cursorInfo.cursor));
-		int caretY = (int) (imblocker$getContentOffsetY() + cursorInfo.cursorLineIndex * 9 - cursorInfo.scrollY);
+		double caretY = imblocker$getContentOffsetY() + cursorInfo.cursorLineIndex * 9 - cursorInfo.scrollY;
 		return new Point(getGuiScale(), caretX, caretY);
 	}
