+++ common/src/main/java/io/github/reserveword/imblocker/mixin/compat/MeteorTextFieldMixin.java
@@ -62,7 +62,7 @@ public abstract class MeteorTextFieldMixin extends MeteorWidgetMixin {
 	@Override
 	public Point getCaretPos() {
 		int caretX = (int) (getTextWidth(cursor) - textStart + pad());
-		return new Point(getGuiScale(), caretX, (int) (height / 3.5));
+		return new Point(getGuiScale(), caretX, height / 3.5);
 	}
 	
 	@Override
